+++ src/transaction/spanner-pubsub/index.ts
@@ -1,0 +1,4 @@
+export { SpannerPubSubTransactionModule } from './module.js';
+export { SpannerPubSubTransactionRunner } from './runner.js';
+export { SpannerStateTransaction } from './state-transaction.js';
+export { SpannerPubSubTransaction } from './transaction.js';