+++ src/transaction/spanner-pubsub/runner.ts
@@ -1,0 +1,84 @@
+import {
+  BufferEventTransaction,
+  TransactionOldTimestampError,
+  TransactionRunner,
+} from '@causa/runtime';
+import { Injectable } from '@nestjs/common';
+import { setTimeout } from 'timers/promises';
+import { PubSubPublisher } from '../../pubsub/index.js';
+import { SpannerEntityManager } from '../../spanner/index.js';
+import { SpannerTransactionOldTimestampError } from './errors.js';
+import { SpannerStateTransaction } from './state-transaction.js';
+import { SpannerPubSubTransaction } from './transaction.js';
+
+/**
+ * The delay, in milliseconds, over which a timestamp issue is deemed irrecoverable.
+ */
+const ACCEPTABLE_PAST_DATE_DELAY = 25000;
+
+/**
+ * A {@link TransactionRunner} that uses Spanner for state and Pub/Sub for events.
+ * A Spanner transaction is used as the main transaction. If it succeeds, events are published to Pub/Sub outside of it.
+ */
+@Injectable()
+export class SpannerPubSubTransactionRunner extends TransactionRunner<SpannerPubSubTransaction> {
+  /**
+   * Creates a new {@link SpannerPubSubTransactionRunner}.
+   *
+   * @param entityManager The {@link SpannerEntityManager} to use for the transaction.
+   * @param publisher The {@link PubSubPublisher} to use for the transaction.
+   */
+  constructor(
+    readonly entityManager: SpannerEntityManager,
+    readonly publisher: PubSubPublisher,
+  ) {
+    super();
+  }
+
+  async run<T>(
+    runFn: (transaction: SpannerPubSubTransaction) => Promise<T>,
+  ): Promise<[T]> {
+    const { result, eventTransaction } = await this.entityManager.transaction(
+      async (dbTransaction) => {
+        const stateTransaction = new SpannerStateTransaction(
+          this.entityManager,
+          dbTransaction,
+        );
+        // This must be inside the Spanner transaction because staged messages should be cleared when the transaction is retried.
+        const eventTransaction = new BufferEventTransaction(this.publisher);
+        const transaction = new SpannerPubSubTransaction(
+          stateTransaction,
+          eventTransaction,
+        );
+
+        try {
+          const result = await runFn(transaction);
+
+          return { result, eventTransaction };
+        } catch (error) {
+          // `TransactionOldTimestampError`s indicate that the transaction is using a timestamp older than what is
+          // observed in the state (Spanner).
+          // Throwing a `SpannerTransactionOldTimestampError` will cause the transaction to be retried with a newer
+          // timestamp.
+          if (!(error instanceof TransactionOldTimestampError)) {
+            throw error;
+          }
+
+          const delay = error.delay ?? Infinity;
+          if (delay >= ACCEPTABLE_PAST_DATE_DELAY) {
+            throw error;
+          }
+          if (delay > 0) {
+            await setTimeout(delay);
+          }
+
+          throw new SpannerTransactionOldTimestampError(error);
+        }
+      },
+    );
+
+    await eventTransaction.commit();
+
+    return [result];
+  }
+}